requirements.txt
attrs==23.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
numpy==2.0.0
pandas==2.2.2
parameterized==0.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
